Moved common styles in StartPage to a shared file
