+++ src/Components/Pages/StartPage/Profile/Styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import {Link} from 'react-router-dom'
+
 export const Table = styled.table`
     width:100%;
     padding: 0.5em;
@@ -18,31 +18,6 @@ export const Th = styled.th`
 
 export const Td = styled.td`
     text-align: left;
-`;
-
-export const Main = styled.div`
-    margin: 0.5em;
-    padding: 0.5em;
-`;
-
-export const TitleButtonContainer = styled.div`
-    display: flex;
-    justify-content: space-between;
-    border-bottom: 1px solid #EDEFF3;
-    padding: 0.5em;
-`;
-
-export const Button = styled(Link)`
-    border-radius: 5px;
-    background-color: rgb(70, 67, 67);
-    color: white;
-    border: none;
-    font-size: 0.9em;
-    padding: 0.5em;
-    text-decoration: none;
-    @media (max-width: 768px) {
-        padding: 0.2em;
-    }
 `;
 
 export const ImageSummaryContainer = styled.div`
